+++ src/app/lib/widgets/boolean-radio/boolean-radio.component.ts
@@ -13,7 +13,13 @@ import {LfbControlWidgetComponent} from '../lfb-control-widget/lfb-control-widge
 export class BooleanRadioComponent  extends LfbControlWidgetComponent implements AfterViewInit {
   static ID = 0;
   _id = BooleanRadioComponent.ID++;
+  options: Map<boolean, string> = new Map([[false, 'No'], [true, 'Yes']]);
+  optionsKeys = []
   ngAfterViewInit() {
+    if(this.formProperty.schema.widget?.optionLabels) {
+      this.options = new Map(this.formProperty.schema.widget.optionLabels);
+    }
+    this.optionsKeys = Array.from(this.options.keys());
     this.labelPosition = 'left';
     super.ngAfterViewInit();
   }